# public/docs/openapi.yaml
openapi: 3.0.1
info:
  title: BiciMarket Payments API
  version: 1.0.0
  description: |
    **Owner**: Rocco Paoloni · **Clerk**: `payments.bicimarket` (admin-only)

    API para gestionar pagos, reembolsos, liquidaciones y payouts.

    **Importante**: Buyers y sellers no se loguean en Payments App. Las vistas administrativas usan JWT con `publicMetadata.admin=true`.
    Los endpoints server-to-server son consumidos por otras apps con `X-Service-Token`.

    **Secciones**:
    - **Pagos**: POST/GET /payments, refund, cancel, confirm (admin override)
    - **Reembolsos**: POST/GET /refunds, detail + status update (admin only)
    - **Comprobantes**: GET /receipts, POST /receipts (Buyer App), detail
    - **Liquidaciones**: GET /settlements, mark_paid (admin), detail
    - **Payouts**: POST/GET /payouts, detail + complete (admin only)
    - **Webhooks**: Mercado Pago + shipment-delivered trigger

  x-api-specs:
    base_path: /api/v1
    webhook_path: /webhooks/mercadopago
    version: v1
    next_version: v2
servers:
  - url: http://localhost:3000
    description: Local Development
    x-env: development
  - url: https://payments.bicimarket.com
    description: Production
    x-env: production
paths:
  /api/v1/payments:
    get:
      summary: Listar pagos
      description: |
        **Filtrar pagos** con paginación estándar.

        Querystring: `?page=1&limit=20&sort=-created_at&orderId=ord_01H…&status=approved`
      tags:
        - "Pagos"
      security:
        - X-Service-Token: []
        - bearerAuth: []
      parameters:
        - in: query
          name: orderId
          schema: { type: string }
          description: Filter by order ID
        - in: query
          name: buyerId
          schema: { type: string }
          description: Filter by buyer profile ID
        - in: query
          name: status
          schema: { type: string, enum: [pending, approved, rejected, cancelled, refunded] }
        - in: query
          name: from
          schema: { type: string, format: date-time }
          description: Filter by start date (ISO 8601)
        - in: query
          name: to
          schema: { type: string, format: date-time }
          description: Filter by end date (ISO 8601)
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 20, maximum: 100 }
      responses:
        '200':
          description: Paginated list of payments
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Payment'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
    post:
      summary: Crear pago
      description: |
        **Crear pago** (Buyer App calls at checkout confirmation).

        Requiere `Idempotency-Key` header. El `items_summary` especifica cómo se splitea el pago entre sellers.
      tags:
        - "Pagos"
      security:
        - X-Service-Token: []
        - bearerAuth: []
      parameters:
        - in: header
          name: Idempotency-Key
          required: true
          schema: { type: string, format: uuid }
          description: UUID for idempotent payment creation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentCreateRequest'
      responses:
        '201':
          description: Payment created successfully with Mercado Pago checkout URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      payment_id:
                        type: string
                        description: Payment ID (pay_...)
                      checkout_url:
                        type: string
                        format: uri
                        description: Mercado Pago checkout URL to redirect the buyer
                      preference_id:
                        type: string
                        description: Mercado Pago preference ID
                  public_key:
                    type: string
                    description: Mercado Pago public key for Wallet SDK
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /api/v1/payments/{paymentId}/refund:
    post:
      summary: Crear reembolso (desde pago)
      description: |
        **Reembolsar parte o todo un pago**.

        Válido después de que el pago sea `approved`. Llamado por Seller App.
        Razones: `seller_rejected`, `buyer_cancelled`, `not_delivered`, `manual`.
      tags:
        - "Pagos"
      security:
        - X-Service-Token: []
      parameters:
        - in: path
          name: paymentId
          required: true
          schema: { type: string, pattern: '^pay_' }
          description: Payment ID (pay_01H...)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefundCreateRequest'
      responses:
        '200':
          description: Refund created and processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Refund'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Clerk JWT token for UI/authenticated endpoints
    X-Service-Token:
      type: apiKey
      in: header
      name: X-Service-Token
      description: Service token for inter-app communication (Buyer, Seller, Shipping)
  schemas:
    Pagination:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        has_more:
          type: boolean
        next_cursor:
          type: string
          nullable: true
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Machine-readable error code
            message:
              type: string
              description: Human-readable error message
            details:
              type: object
              nullable: true
              description: Additional error details
    Payment:
      type: object
      properties:
        id:
          type: string
          description: Payment ID (pay_...)
        order_id:
          type: string
          description: Order ID (ord_...)
        buyer_profile_id:
          type: string
          description: Buyer profile ID (byp_...)
        buyer_clerk_user_id:
          type: string
          nullable: true
        amount_cents:
          type: integer
          description: Amount in cents
        currency:
          type: string
          default: ARS
        status:
          type: string
          enum: [pending, approved, rejected, cancelled, refunded]
        method:
          type: string
          nullable: true
          description: Payment method (credit_card, debit_card, etc)
        card_last4:
          type: string
          nullable: true
        gateway_reference:
          type: string
          nullable: true
          description: Mercado Pago payment ID
        approved_at:
          type: string
          format: date-time
          nullable: true
        rejected_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, order_id, buyer_profile_id, amount_cents, currency, status, created_at, updated_at]
    PaymentCreateRequest:
      type: object
      properties:
        order_id:
          type: string
          description: Order ID
        buyer_clerk_user_id:
          type: string
          description: Buyer Clerk user ID
        buyer_profile_id:
          type: string
          description: Buyer profile ID
        buyer_email:
          type: string
          format: email
          description: Buyer email for MP payer
        amount_cents:
          type: integer
          description: Total amount in cents
        currency:
          type: string
          default: ARS
        items_summary:
          type: array
          description: Breakdown by seller
          items:
            type: object
            properties:
              seller_profile_id:
                type: string
              subtotal_cents:
                type: integer
              shipping_cost_cents:
                type: integer
            required: [seller_profile_id, subtotal_cents, shipping_cost_cents]
        return_urls:
          type: object
          properties:
            success:
              type: string
            failure:
              type: string
            pending:
              type: string
      required: [order_id, buyer_clerk_user_id, buyer_profile_id, amount_cents]
    Refund:
      type: object
      properties:
        id:
          type: string
          description: Refund ID (ref_...)
        payment_id:
          type: string
        amount_cents:
          type: integer
        currency:
          type: string
          default: ARS
        status:
          type: string
          enum: [pending, approved, rejected, failed]
        reason:
          type: string
          description: Reason code (seller_rejected, buyer_cancelled, not_delivered, manual)
        gateway_reference:
          type: string
          nullable: true
          description: Mercado Pago refund ID
        seller_profile_id:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    RefundCreateRequest:
      type: object
      properties:
        amount_cents:
          type: integer
          description: Refund amount in cents
        reason:
          type: string
          enum: [seller_rejected, buyer_cancelled, not_delivered, manual]
        seller_profile_id:
          type: string
          nullable: true
      required: [amount_cents, reason]
...
  responses:
    BadRequest:
      description: Invalid request payload
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing or invalid authentication (JWT or service token)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Invalid state transition or business logic conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'